Add: base_fus_updater.bbclass
This class implements two tasks:
* do_create_nand_packages:
This task is started after do_rootfs and creates the nand partitions for
the nand version. You can identify them with searching for the suffix
*.nand.* inside the image output directory.

* do_create_update_package:
Create the rauc update packge for the nand and the emmc version. The
task is started after do_image_wic. For the nand version the right *.dtb
(device tree) is identified with UBOOT_DTB_NAME. The images are placed
into the standard image output directory