Add: conf/machine/fsimx8mm-extra.conf
Add the extension of the machine description. For each machine that
should support the update concept, you have to create a new *-extra.conf
and include it inside the conf/layer.conf. Instead you can generalize
the machine description and use one for all. Currently only the fsimx8mm
is compatible and the machine file of meta-fus must be prepared for
extension (e.g. IMAGE_FSTYPES ?= ... ).