Fix: Permission problem
Currently the permissions of the rootfilesystem are not set correctly.
The permission of the developing system are used. You have to use the
IMAGE_FSTYPES interface. To set the image generation task as independet
task, the permissions will not correct set.

You have to use this construction:
e.g. update_package:

""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
IMAGE_CMD_update_package () {
	do_create_squashfs_rootfs_images # the former image creation
task
}
""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""