Remove: MEMORY_TYPE
This removes the distinction between emmc and nand for the rootfs.
This commit goes along with the "Add: soft link to config files" commit in the dynamic_overlay git.
Both files are now installed on the board, and the dynamic overlay creates a soft link to the file fitting the board config.
Add python version of FS-Update until the C++ version is complete
Move FIRMWARE_VERSION into the rauc recipe