classes:cert-handler: always do_generate_certificates
Certificates are not stored in SSTATE as installed files of a package,
but reside in the recipe's meta-layer.

If TMPDIR is regenerated from SSTATE, the certificates in the meta-layer
might be missing, even though do_generate_certificates is already marked
as executed in the cache.

Since do_generate_certificates cannot overwrite certificates, it should
always be executed during the build for stability reasons.